Major changes to improve usability and extendability.
This is done to unify some internal interfaces, and to
prepare for some major feature additions.

Split all methods which simply manipulate point coordinates
into a module named 'patternmath'. These methods manipulate points,
and shouldn't require any knowledge of pattern class objects.

Created a simple class dPnt(), which is a light weight stand-in
for the Point class, having only x and x attributes.

Add a helper to make the addition of patterns and pieces in the design
file more sensible and easier to understand for the designer.

Changed most class initializers and methods so that when a point
(location) is required, they will accept either an object with
x and y attributes (typically a point object), or an x,y tuple.

Added addPoint method to pattern piece, instead of creating points
and adding them later.

Added helpers to pattern piece class to add things like
grainline, cutting line, etc. This reduces the number of lines
required in the pattern file.

Added control points as optional sub-points to class Point()
These are 'inpoints' and 'outpoints', and are used in drawing
SVG paths, for Bezier control points.

Removed a lot of cruft and unused methods and classes.

Modified the Allington sample pattern to use the new methods.

Signed-off-by: Steve Conklin <sconklin@canonical.com>